requirements.txt
alembic==1.15.1
aniso8601==10.0.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.42.2
colorama==0.4.6
cryptography==44.0.2
Flask==3.1.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
ngrok==1.4.0
psycopg2-binary==2.9.10
pycparser==2.22
PyMySQL==1.1.1
pyngrok==7.2.3
python-dotenv==1.1.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.1.2
WTForms-SQLAlchemy==0.4.2
